Develop (#55)
* remove bulk of sys_conf_wrap.c along with superfluous references

* remove external.gi

* now longer lookup "path_sep"

* use GetEnv to set PATH_SEP and use instead of looking up "path_sep"

* remove/replace all SWIG-based lookup functions

* remove sys_conf_wrap.c

* add functions in sys_conf.g to AUTO list

* remove explicit RandomSeed(TimeInSecs()) and return to default RandomSeed(1) as
documented in GAP manual.  This allows actions that use Random()/RandomList() to
be repeatable.

* remove obsolete auxiliary files

* remove sys_conf.g

* move SyPathSep() to GapUtils.c and rename to PathSep()

* replace SPIRAL_CONFIG_PATH_SEP environment variable with PathSep() function

* replace SPIRAL_CONFIG_SPIRAL_DIR with SPIRAL_DIR

* remove fpgen

* remove installer

* replace three similar functions sys_exists(), FileExists(), and CheckFileExists()
with single new FileExists()

* remove quotify stuff

* remove unused error messages and error types

* remove obsolete TmpName() and supporting chain of functions

* remove spiral_spl_prog.c/h and move ReadValFromFile() to spiral.c
remove GuSysCheckExists()
remove SetEnv() and supporting functions

* remove obsolete functions: config_demand_val(), remove config_demand_val_profile(), remove config_valid_strval_profile(), config_valid_strval(), remove config_get_val(), remove sys_mkdir()

* remove leading space in registration name for ReadVal

* include read.h to define ReadIt()
