+++ config/spirald.in
@@ -17,9 +17,8 @@ if [ ! -f "$GAP_EXE" ]; then
 fi
 
 export SPIRAL_CONFIG_TMP_DIR="/tmp/spiraltmp"
-export SPIRAL_CONFIG_SPIRAL_DIR="$SPIRAL"
+export SPIRAL_DIR="$SPIRAL"
 export SPIRAL_CONFIG_EXEC_DIR="/usr/local/bin"
-export SPIRAL_CONFIG_PATH_SEP="/"
 
 ##  Setup PROFILER_LOCAL_ARGS with the target used for CMeasure & CMatrix measurements
 export PROFILER_LOCAL_ARGS="--target @PROFILER_TARGET@"
