--- gap/bin/gap-manual.in
@@ -1,2 +1,0 @@
-#!/bin/sh
-xdvi -s 8 @spiraldir@/gap/doc/manual.dvi &
