+++ gap/lib/init.g
@@ -132,13 +132,9 @@ function(arg)
     ClearCommentBuffer();
 end);
 
-##  CVSID := function(cvs_id)
-##      CurrentFile().id := cvs_id;
-##  end;
 
-VERLIB := "v3r4p4 1997/04/18";
-
-PATH_SEP := config_val_t_strval_get(config_demand_val("path_sep"));
+PATH_SEP := PathSep();
+SPIRAL_DIR := GetEnv("SPIRAL_DIR");
 
 Global.SpiralVersion := Version();
 
@@ -313,9 +309,6 @@ AUTO( ReadLib( "abattoir" ),
   Save, SetPkgname, PKGNAME, LOADED_PACKAGES, ReadPkg, ExecPkg, LoadPackage,
   RequirePackage, IsOperationsRecord, OpsOps, OperationsRecord, EXEC,
   False, True, PrintFactorsInt );
-
-AUTO( ReadLib("sys_conf"),
-    Conf );
 
 AUTO( ReadLib( "agcent" ),
   MainEntryCSAgGroup, CentralCaseCentAgGroup, GeneralCaseCentAgGroup,
@@ -1046,39 +1039,6 @@ AUTO( ReadGrp( "trans" ),
   SetSetOrbits, OrbNEq, CntOp, TransitiveIdentification,
   SelectTransitiveGroups, AllTransitiveGroups, OneTransitiveGroup );
 
-#AUTO( ReadTwo( "twogp" ),
-#  TGParts, TGLoad, TGGroup, TwoGroup, pClass, AllTwoGroups, OneTwoGroup );
-#
-#AUTO( ReadThr( "thrgp" ),
-#  ThGParts, ThGLoad, ThGGroup, ThreeGroup, AllThreeGroups, OneThreeGroup );
-#
-#AUTO( ReadTbl( "ctadmin" ),
-#  TABLEFILENAME, LIBTABLE, SET_TABLEFILENAME, GALOIS, TENSOR, EvalChars, MBT,
-#  MOT, LowercaseString, NotifyCharTableName, NotifyCharTable,
-#  LibInfoCharTable, FirstNameCharTable, FileNameCharTable, ALN, ALF, ACM, ARC,
-#  ConstructMixed, ConstructProj, ConstructDirectProduct, ConstructIsoclinic,
-#  ConstructV4G, ConstructGS3, ConstructPermuted, ConstructSubdirect,
-#  UnpackedCll, CllToClf, ConstructClifford, BrauerTree, DecMat,
-#  BasicSetBrauerTree, AddDecMats, PartsBrauerTableName, BrauerTable,
-#  LibraryTables, CharTableLibrary, OfThose, IsSporadicSimple, SchurCover,
-#  AllCharTableNames, ShrinkClifford, TextString, BlanklessPrint, ShrinkChars,
-#  ClfToCll, PrintFusion, PrintToLib, PrintClmsToLib );
-#
-#AUTO( ReadTbl( "ctprimar" ),
-#  LIBLIST );
-#
-#AUTO( ReadTom( "tmprimar" ),
-#  TOM, TOMLIST, TomLibrary );
-#
-#AUTO( ReadSml( "idgroup.sml" ),
-#  AgGroupCode, CodeAgGroup, InitRandomIsomorphismChecking, RandomSpecialPres,
-#  RandomIsomorphismChecking, IdGroupRandomTest, IdGroupSpecialFp, EvalFpCoc,
-#  IdSmallGroup, IdP1Q1R1Group, IdP2Q1Group, IdP1Q2Group, IdP1Group, IdP2Group,
-#  IdP3Group, IdP1Q1Group, IdGroup );
-#
-#AUTO( ReadSml( "smallgrp.sml" ),
-#  Codes1000, PermGroupCode, AgGroupCode, GroupCode, LoadSmallGroups,
-#  UnloadSmallGroups, SmallGroup, AllSmallGroups, NumberSmallGroups );
 
 # load packages
 
@@ -1104,8 +1064,6 @@ CantCopy(Cyclotomics);
 CantCopy(Integers);
 
 Concat:=ConcatenationString;
-QUOTIFY := file_quotify_static;
-RandomSeed(TimeInSecs());
 
 # Info Lattice
 # ------------
@@ -1114,6 +1072,6 @@ if not IsBound(InfoLatticeSpiral) then
   InfoLatticeSpiral := Ignore;
 fi;
 
-Read(Concat(Conf("spiral_dir"), Conf("path_sep"), "namespaces", Conf("path_sep"), "init.g"));
+Read(Concat(SPIRAL_DIR, PATH_SEP, "namespaces", PATH_SEP, "init.g"));
 
 _fileProgress.done := true;
