--- gap/src/spiral_spl_prog.h
@@ -1,6 +1,0 @@
-/****************************************************************************
-**
-*F  InitSPIRAL_SPLProg() . . . . . . . . . . . initialize SPL program package
-**
-*/
-void            InitSPIRAL_SPLProg ( );
