+++ gap/src/system.h
@@ -369,12 +369,6 @@ extern  int             SyExec ( char * cmd );
 extern  UInt   SyTime ( void );
 
 
-/****************************************************************************
-**
-*F  SyTmpname() . . . . . . . . . . . . . . . . . return a temporary filename
-*/
-extern  char *          SyTmpname ( void );
-
 
 /****************************************************************************
 **
@@ -665,7 +659,6 @@ typedef void            (*TNumAbortFuncBags) (
     Char* msg);
 
 extern  TNumAbortFuncBags       AbortFuncBags;
-
 
 #endif // _SYSTEM_H
