+++ namespaces/spiral/defaults.gi
@@ -209,13 +209,13 @@ IBMC99Mixin := rec(
 
 #  NOTE: I'd like to pull that out into a function call, but failed to use load/include/read inside a function... => ask YSV
 #if LocalConfig.osinfo.isWindows() then
-#    Exec(let(sdir:=Conf("spiral_dir"), Concat("SubWCRev.exe ", sdir, " ", Concat(sdir, "\\spiral\\svn_win.src ", sdir, "\\spiral\\svn_info.g > NUL"))));
+#    Exec(let(sdir:=SPIRAL_DIR, Concat("SubWCRev.exe ", sdir, " ", Concat(sdir, "\\spiral\\svn_win.src ", sdir, "\\spiral\\svn_info.g > NUL"))));
 #    Load(svn_info);
 #fi;
 #if LocalConfig.osinfo.isLinux() then
 # NOTE: For now, assume that any non-windows system is a linux system.
 #else
- #   Exec(let(sdir:=Conf("spiral_dir"), Concat(". ", sdir, "/spiral/svn_linux.src ", sdir, " > ", sdir, "/spiral/svn_info.g")));
+ #   Exec(let(sdir:=SPIRAL_DIR, Concat(". ", sdir, "/spiral/svn_linux.src ", sdir, " > ", sdir, "/spiral/svn_info.g")));
 #    Load(svn_info);
 #fi;
